Rewrote entire bot for readability and expansion

+++ utilities/database.py
@@ -1,22 +1,86 @@
 import sqlite3, os
 
+'''
+ADD TRACKING UTILITY FUNCTIONS
 
-def add_channels(channel_id):
+add_suggestion_channel - adds tracking for suggestions to the given channel_id
+
+add_recommendation_channel - adds tracking for recommendations to the given channel_id
+'''
+
+def add_suggestion_channel(channel_id):
     conn = sqlite3.connect("chh.db")
-    conn.execute("INSERT INTO allowed (ID) VALUES ({})".format(channel_id))
+    conn.execute('''CREATE TABLE IF NOT EXISTS allowed
+                    (ID INT PRIMARY KEY NOT NULL)''')
     conn.commit()
-    conn.close()
+    cursor = conn.execute("SELECT * FROM allowed WHERE ID={}".format(channel_id))
+    data = cursor.fetchall()
+    if len(data) == 0:
+        conn.execute("INSERT INTO allowed (ID) VALUES ({})".format(channel_id))
+        conn.commit()
+        conn.close()
+        return True
+    else:
+        return False
 
-def remove_channel(channel_id):
+def add_recommendation_channel(channel_id):
     conn = sqlite3.connect("chh.db")
-    conn.execute("DELETE FROM allowed WHERE ID = {}".format(channel_id))
+    conn.execute('''CREATE TABLE IF NOT EXISTS allowed_recommended
+                    (ID INT PRIMARY KEY NOT NULL)''')
     conn.commit()
-    conn.close()
+    cursor = conn.execute("SELECT * FROM allowed_recommended WHERE ID={}".format(channel_id))
+    data = cursor.fetchall()
+    if len(data) == 0:
+        conn.execute("INSERT INTO allowed_recommended (ID) VALUES ({})".format(channel_id))
+        conn.commit()
+        conn.close()
+        return True
+    else:
+        return False
 
 
+'''
+REMOVE TRACKING UTILITY FUNCTIONS
 
-def get_allowed_channels():
+remove_suggestion_channel - removes tracking for suggestions from the channel with given ID
+
+remove_recommendation_channel - removes tracking for recommendations from the channel with given id
+'''
+def remove_suggestion_channel(channel_id):
     conn = sqlite3.connect("chh.db")
+    conn.execute('''CREATE TABLE IF NOT EXISTS allowed
+                    (ID INT PRIMARY KEY NOT NULL)''')
+    conn.commit()
+    cursor = conn.execute("SELECT * FROM allowed WHERE ID={}".format(channel_id))
+    data = cursor.fetchall()
+    if len(data) == 0:
+        return False
+    else:
+        conn.execute("DELETE FROM allowed WHERE ID = {}".format(channel_id))
+        conn.commit()
+        conn.close()
+        return True
+
+def remove_recommendation_channel(channel_id):
+    conn = sqlite3.connect("chh.db")
+    conn.execute('''CREATE TABLE IF NOT EXISTS allowed_recommended
+                    (ID INT PRIMARY KEY NOT NULL)''')
+    conn.commit()
+    cursor = conn.execute("SELECT * FROM allowed_recommended WHERE ID={}".format(channel_id))
+    data = cursor.fetchall()
+    if len(data) == 0:
+        return False
+    else:
+        conn.execute("DELETE FROM allowed_recommended WHERE ID = {}".format(channel_id))
+        conn.commit()
+        conn.close()
+        return True
+
+def get_suggestion_channels():
+    conn = sqlite3.connect("chh.db")
+    conn.execute('''CREATE TABLE IF NOT EXISTS allowed
+                    (ID INT PRIMARY KEY NOT NULL)''')
+    conn.commit()
     cursor = conn.execute("SELECT * FROM allowed").fetchall()
     ids = []
     for row in cursor:
@@ -24,22 +88,7 @@ def get_allowed_channels():
     conn.close()
     return ids
 
-def add_recommended_channel(channel_id):
-    conn = sqlite3.connect("chh.db")
-    conn.execute('''CREATE TABLE IF NOT EXISTS allowed_recommended
-                    (ID INT PRIMARY KEY NOT NULL)''')
-    conn.commit()
-    conn.execute("INSERT INTO allowed_recommended (ID) VALUES ({})".format(channel_id))
-    conn.commit()
-    conn.close()
-
-def remove_recommended_channel(channel_id):
-    conn = sqlite3.connect("chh.db")
-    conn.execute("DELETE FROM allowed_recommended WHERE ID = {}".format(channel_id))
-    conn.commit()
-    conn.close()
-
-def get_allowed_recommended_channels():
+def get_recommended_channels():
     conn = sqlite3.connect("chh.db")
     conn.execute('''CREATE TABLE IF NOT EXISTS allowed_recommended
                     (ID INT PRIMARY KEY NOT NULL)''')
